+++ src/ui/components/layer-file-upload-button.js
@@ -52,7 +52,15 @@ registerComponent('layer-file-upload-button', {
    * @param {Layer.Core.MessageTypeModel[]} evt.detail.models
    */
   events: ['layer-files-selected', 'layer-models-generated'],
-  template: '<label layer-id="label">+</label><input layer-id="input" type="file"></input>',
+  template: `
+    <label layer-id="label">
+      <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48">
+        <path d="M0 0h48v48H0z" fill="none"/>
+        <path class="layer-file-upload-plus" d="M26 14h-4v8h-8v4h8v8h4v-8h8v-4h-8v-8zM24 4C12.95 4 4 12.95 4 24s8.95 20 20 20 20-8.95 20-20S35.05 4 24 4zm0 36c-8.82 0-16-7.18-16-16S15.18 8 24 8s16 7.18 16 16-7.18 16-16 16z"/>
+      </svg>
+    </label>
+    <input layer-id="input" type="file"></input>
+  `,
   style: `
     layer-file-upload-button {
       cursor: pointer;